Re-set comment approval status for always allowed domains
Even after the `comment_approved` key is set to `1` in the comment
data, the comment goes through `wp_new_comment()` for storage. This
in turn runs the comment through `wp_allow_comment()` which causes
the `comment_approved` key to be set back to `0` if any kind of
comment moderation is enabled on the site.

This uses the avaialble `pre_comment_approved` filter to check if
the comment is a webmention previously approved based on its
domain name and then returns `1` as expected.